+++ src/fi/testbed2/exception/DownloadTaskException.java
@@ -4,10 +4,6 @@ public class DownloadTaskException extends Exception {
 
 	private static final long serialVersionUID = -1L;
 
-	public DownloadTaskException() {
-		super();
-	}
-
 	public DownloadTaskException(String detailMessage, Throwable throwable) {
 		super(detailMessage, throwable);
 	}
@@ -16,9 +12,4 @@ public class DownloadTaskException extends Exception {
 		super(detailMessage);
 	}
 
-	public DownloadTaskException(Throwable throwable) {
-		super(throwable);
-	}
-
-	
 }